correction bugs diagonnales

+++ src/Bouton_rond.java
@@ -14,7 +14,6 @@ public class Bouton_rond extends JButton{
 
     protected void paintComponent(Graphics g) {
         g.setColor(c);
-        g.fillOval((int) (this.getWidth() / 6),(int) (this.getHeight() / 6),(int) (this.getWidth() / 1.5),(int) (this.getHeight() / 1.5));
+        g.fillOval(this.getWidth() / 6, this.getHeight() / 6,(int) (this.getWidth() / 1.5),(int) (this.getHeight() / 1.5));
     }
-
 }
